backend/pyproject.toml
[project]
name = "HuggingSpace-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dj-database-url==2.1.0",
    "django==5.0.1",
    "django-cors-headers==4.3.1",
    "django-environ==0.11.2",
    "django-guardian==2.4.0",
    "djangorestframework==3.14.0",
    "djangorestframework-simplejwt==5.3.1",
    "drf-spectacular==0.27.1",
    "psycopg2-binary==2.9.9",
    "pytest==7.4.4",
    "pytest-cov==4.1.0",
    "pytest-django==4.7.0",
    "python-dotenv==1.0.0",
    "python-jose==3.3.0",
    "requests==2.31.0",
    "supabase-py==2.3.4",
]
